Main Nav setup. ONto pages...

+++ src/RommStar.Core/Services/LoggingService.cs
@@ -40,7 +40,7 @@ namespace RommStar.Core.Services
 
         public void Log(string message, LoggingLevel logLevel = LoggingLevel.Normal)
         {
-            if (logLevel > Settings.Default.LoggingLevel)
+            if ((int)logLevel > Settings.Default.LoggingLevel)
             {
                 return; // Skip logging if the message's log level is higher than the configured log level
             }
